% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genes_annot}
\alias{genes_annot}
\title{genes_annot data set documentation}
\format{
An object of class \code{data.frame} with 19100 rows and 6 columns.
}
\usage{
genes_annot
}
\description{
genes_annot description goes here
}
\keyword{datasets}
