% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE112679_genes_annot}
\alias{GSE112679_genes_annot}
\title{A data frame describing the features}
\format{
genes_annot is a data frame with 19100 rows and 6 variables:
\describe{
\item{Symbol}{  chr  "OR4F5" "AL627309.1" "OR4F29" "OR4F16" ...}
\item{Chr}{     chr  "chr01" "chr01" "chr01" "chr01" ...}
\item{Strand}{  chr  "+" "-" "+" "-" ...}
\item{Start}{   num  69091 134901 367640 621096 ...}
\item{Stop}{    num  70008 139379 368634 622034 ...}
\item{Length}{  num  918 4479 995 939 606 ...}
}
}
\usage{
GSE112679_genes_annot
}
\description{
Data frame with 19100 rows and 6 variables describing features
}
\keyword{datasets}
