% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE112679_Val2_featureCount}
\alias{GSE112679_Val2_featureCount}
\title{A matrix of feature counts for the samples in validation set 2}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 19100 rows and 240 columns.
}
\usage{
GSE112679_Val2_featureCount
}
\description{
GSE112679_Val2_featureCount - a matrix of feature counts for the samples in validation set 2
rows are feature; columns are samples.
}
\keyword{datasets}
