% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Train_featureCount}
\alias{Train_featureCount}
\title{Train_featureCount data set documentation}
\format{
An object of class \code{data.frame} with 19100 rows and 1120 columns.
}
\usage{
Train_featureCount
}
\description{
Train_featureCount description goes here
}
\keyword{datasets}
