% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TUTI_featureCount}
\alias{TUTI_featureCount}
\title{A matrix of feature counts for the matched Tumor/Tissue samples}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 19100 rows and 52 columns.
}
\usage{
TUTI_featureCount
}
\description{
TUTI_featureCount - a matrix of feature counts for the matched Tumor/Tissue samples
rows are feature; columns are samples.
}
\keyword{datasets}
