% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Val1_featureCount}
\alias{Val1_featureCount}
\title{Val1_featureCount data set documentation}
\format{
An object of class \code{data.frame} with 19100 rows and 1194 columns.
}
\usage{
Val1_featureCount
}
\description{
Val1_featureCount description goes here
}
\keyword{datasets}
