% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Val2_featureCount}
\alias{Val2_featureCount}
\title{Val2_featureCount data set documentation}
\format{
An object of class \code{data.frame} with 19100 rows and 240 columns.
}
\usage{
Val2_featureCount
}
\description{
Val2_featureCount description goes here
}
\keyword{datasets}
