% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Train_featureCount}
\alias{Train_featureCount}
\title{A matrix of feature counts for the samples in the training set}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 19100 rows and 1120 columns.
}
\usage{
Train_featureCount
}
\description{
Train_featureCount - a matrix of feature counts for the samples in the training set
rows are feature; columns are samples.
}
\keyword{datasets}
