% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TUTI_featureCount}
\alias{TUTI_featureCount}
\title{TUTI_featureCount data set documentation}
\format{
An object of class \code{data.frame} with 19100 rows and 52 columns.
}
\usage{
TUTI_featureCount
}
\description{
TUTI_featureCount description goes here
}
\keyword{datasets}
