% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Val1_featureCount}
\alias{Val1_featureCount}
\title{A matrix of feature counts for the samples in validation set 1}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 19100 rows and 1194 columns.
}
\usage{
Val1_featureCount
}
\description{
Val1_featureCount - a matrix of feature counts for the samples in validation set 1
rows are feature; columns are samples.
}
\keyword{datasets}
